Modify how magic is found and read movie header
The way we used to find magic was prone for error, and it hadn't failed
because of a coincidence---namely, "dumb header\n" is 12 characters
long, which is a multiple of the length of the magic word "TemI".

I now find the magic word in a much more reasonable way. Incidently,
everything up to magic is the movie header, which we had previously
never read.